Issue #7: Implement logging to file Issue #22: Extend testrunner to support extra video params Issue #23: Write a script to batch produce sample MPEG files
We now have testrunner logging to a configurable directory. CSV files will be created for each call to testrunner.
The files will have the date and time as their name. Inside the csv files are the results of each stream called, with all relevant parameters noted.
Testrunner also uses functions to be better readable. It now calls each server for each configured format using nested loops.

tcpdump now has a configurable outfile for stderr, this is used to gather the number of packets quickly.

makempeg now works and is reasonably configurable. It expects stream.config in the path. It uses the same arrays as testrunner to create the videos.