Issue #24: Expand the controllsrv UI for testruns
Implemented the reporting of ffmpeg memory usage to the controlsrv ui.

Also set and remove a simple flag during each test run. Then the controlsrv JSONstats will check for the presence of this flag and report whether a test is in process. This way we report something more of interest to the user.